move form logic to another file
